# Prepare for Go 1.9 math/bits package.