###### Required for all non-trivial PRs
- [x] Rebased/mergable
- [x] Tests pass
- [x] CHANGELOG.md updated

**Note** Should not be merged until we decide to move to Go 1.9.

This PR replaces the `github.com/dgryksi/go-bits` dependency with the standard library `math/bits` package.