@e-dard 2017-08-25
@jwilder sorry I did run the benchmarks before pushing this, but didn't attach them to the issue. Here are the ones for the `Float{Encoder, Decoder}`:

```
⇒  benchstat /tmp/old.txt /tmp/new.txt
name            old time/op    new time/op    delta
FloatEncoder-4    7.15µs ±10%    6.95µs ± 6%    ~           (p=0.119 n=11+10)
FloatDecoder-4    1.89µs ± 7%    1.81µs ± 3%  -4.23%        (p=0.003 n=11+11)

name            old alloc/op   new alloc/op   delta
FloatEncoder-4      624B ± 0%      624B ± 0%    ~     (all samples are equal)
FloatDecoder-4    0.00B ±NaN%    0.00B ±NaN%    ~     (all samples are equal)

name            old allocs/op  new allocs/op  delta
FloatEncoder-4      4.00 ± 0%      4.00 ± 0%    ~     (all samples are equal)
FloatDecoder-4     0.00 ±NaN%     0.00 ±NaN%    ~     (all samples are equal)
```

Not a huge amount in it; I think the change is fine.

I'll add build tags.