@e-dard 2017-09-26
@jwilder I think this is in good shape now. Had to do some fiddly abstractions to ensure that the committed code worked both on Go 1.9 and previous versions. When we commit to Go 1.9 I'll remove the `pkg/bits` package and directly call into `math/bits` from the `hll` and `tsm1`.